Created initial Page Application feature code, news page application dummy path generator (#3082); Global Fixture test created which drops/creates model and builds initial pages/authorization data; Fixed sha1 Doctrine custom type from messing the schema updates.
git-svn-id: http://svn.vig/supra7/trunk@466 f05b830f-2085-0410-a479-8f75d39d0b34

+++ src/lib/Supra/Controller/Pages/Request/PageRequestEdit.php
@@ -160,10 +160,13 @@ class PageRequestEdit extends PageRequest
 		// 10. Clear all property metadata on the public going to be merged
 		//TODO: remove reference elements as well!
 		$propertyIdList = $draftProperties->collectIds();
-		$qb = $publicEm->createQueryBuilder();
-		$qb->delete(Entity\BlockPropertyMetadata::CN(), 'r')
-				->where($qb->expr()->in('r.blockProperty', $propertyIdList))
-				->getQuery()->execute();
+		
+		if ( ! empty($propertyIdList)) {
+			$qb = $publicEm->createQueryBuilder();
+			$qb->delete(Entity\BlockPropertyMetadata::CN(), 'r')
+					->where($qb->expr()->in('r.blockProperty', $propertyIdList))
+					->getQuery()->execute();
+		}
 		
 		// 11. Merge all properties from 5
 		foreach ($draftProperties as $property) {
